Thêm CRUD: cho ChuyenKhoan

+++ QuanLyNganHang/ET/ET_ChuyenKhoan.cs
@@ -9,11 +9,11 @@ namespace ET
 
         public string MACK { get; set; }
         public DateTime NGAYCK { get; set; }
-        public float SOTIEN { get; set; }
+        public decimal SOTIEN { get; set; }
         public string MATKGUI { get; set; }
         public string MATKNHAN { get; set; }
         public string NOIDUNG { get; set; }
-        public ET_ChuyenKhoan(string mACK, DateTime nGAYCK, float sOTIEN, string mATKGUI, string mATKNHAN, string nOIDUNG)
+        public ET_ChuyenKhoan(string mACK, DateTime nGAYCK, decimal sOTIEN, string mATKGUI, string mATKNHAN, string nOIDUNG)
         {
             MACK = mACK;
             NGAYCK = nGAYCK;
